move module my page screen
