+++ app/src/main/java/com/jin/honey/feature/navigation/RootNavigation.kt
@@ -61,8 +61,8 @@ import com.jin.ui.foodsearch.FoodSearchScreen
 import com.jin.ui.foodsearch.FoodSearchViewModel
 import com.jin.ui.ingredient.IngredientScreen
 import com.jin.ui.ingredient.IngredientViewModel
-import com.jin.honey.feature.mypage.ui.MyPageScreen
-import com.jin.honey.feature.mypage.ui.MyPageViewModel
+import com.jin.ui.mypage.MyPageScreen
+import com.jin.ui.mypage.MyPageViewModel
 import com.jin.honey.feature.onboarding.ui.OnboardingScreen
 import com.jin.honey.feature.onboarding.ui.OnboardingViewModel
 import com.jin.domain.chat.ChatRepository
